+++ arrow-udf-wasm/src/lib.rs
@@ -34,6 +34,8 @@ pub struct Runtime {
     config: Config,
     /// Function names.
     functions: HashSet<String>,
+    /// User-defined types.
+    types: HashMap<String, String>,
     /// Instance pool.
     instances: Mutex<Vec<Instance>>,
 }
@@ -66,6 +68,7 @@ impl Debug for Runtime {
         f.debug_struct("Runtime")
             .field("config", &self.config)
             .field("functions", &self.functions)
+            .field("types", &self.types)
             .field("instances", &self.instances.lock().unwrap().len())
             .finish()
     }
@@ -96,21 +99,26 @@ impl Runtime {
             .find_map(|e| e.name().strip_prefix("ARROWUDF_VERSION_"))
             .context("version not found")?;
         let (major, minor) = version.split_once('_').context("invalid version")?;
-        ensure!(major == "1", "unsupported abi version: {major}.{minor}");
+        ensure!(major <= "2", "unsupported abi version: {major}.{minor}");
 
         let mut functions = HashSet::new();
+        let mut types = HashMap::new();
         for export in module.exports() {
-            let Some(encoded) = export.name().strip_prefix("arrowudf_") else {
-                continue;
-            };
-            let name = base64_decode(encoded).context("invalid symbol")?;
-            functions.insert(name);
+            if let Some(encoded) = export.name().strip_prefix("arrowudf_") {
+                let name = base64_decode(encoded).context("invalid symbol")?;
+                functions.insert(name);
+            } else if let Some(encoded) = export.name().strip_prefix("arrowudt_") {
+                let meta = base64_decode(encoded).context("invalid symbol")?;
+                let (name, fields) = meta.split_once('=').context("invalid type string")?;
+                types.insert(name.to_string(), fields.to_string());
+            }
         }
 
         Ok(Self {
             module,
             config,
             functions,
+            types,
             instances: Mutex::new(vec![]),
         })
     }
@@ -118,6 +126,11 @@ impl Runtime {
     /// Return available functions.
     pub fn functions(&self) -> impl Iterator<Item = &str> {
         self.functions.iter().map(|s| s.as_str())
+    }
+
+    /// Return available types.
+    pub fn types(&self) -> impl Iterator<Item = (&str, &str)> {
+        self.types.iter().map(|(k, v)| (k.as_str(), v.as_str()))
     }
 
     /// Call a function.